[NFC] move index_utils to _support.indexing
