+++ iree/turbine/kernel/wave/codegen/read_write.py
@@ -41,9 +41,14 @@ from ...compiler.vector_codegen import (
 from ...ops.wave_ops import get_custom, read, write, CustomOp
 
 from ..utils.general_utils import find_index_bounds, get_fastest_index
-from ..utils.symbol_utils import safe_subs, subs_idxc
 
-from ..._support.indexing import IndexingContext, IndexExpr, IndexSequence, index_symbol
+from ..._support.indexing import (
+    IndexingContext,
+    IndexExpr,
+    IndexSequence,
+    safe_subs,
+    subs_idxc,
+)
 from ...lang.wave_types import IndexMapping
 from ...lang.global_symbols import *
