+++ iree/turbine/kernel/wave/global_to_shared_gathers.py
@@ -11,7 +11,7 @@ from ..wave.constraints import (
     TilingConstraint,
 )
 from .._support.tracing import CapturedTrace
-from .._support.indexing import IndexSequence, IndexSymbol, IndexExpr
+from .._support.indexing import IndexSequence, IndexSymbol, IndexExpr, subs_idxc
 from ..lang.wave_types import IndexMapping
 from ..ops.wave_ops import Read, Write, get_custom
 from ..lang.global_symbols import *
@@ -19,7 +19,6 @@ from math import prod
 import torch.fx as fx
 from collections import defaultdict
 from copy import deepcopy
-from .utils.symbol_utils import subs_idxc
 from .utils.general_utils import is_gather
 from .minimize_global_loads import (
     has_write_shared_user,
