+++ iree/turbine/kernel/wave/scheduling/graph_utils.py
@@ -7,13 +7,12 @@
 import torch.fx as fx
 from random import Random
 from collections import defaultdict, deque
-from ..._support.indexing import index_symbol, IndexExpr
+from ..._support.indexing import index_symbol, IndexExpr, safe_subs
 from .resources import *
 from dataclasses import dataclass
 import sympy
 import math
 from functools import partial
-from ..utils.symbol_utils import safe_subs
 import multiprocessing as mp
 from typing import Optional, Callable
