+++ iree/turbine/kernel/wave/scheduling/loop_reconstruction_utils.py
@@ -11,7 +11,6 @@ from ...ops.wave_ops import (
     Placeholder,
 )
 from ..utils.graph_utils import graph_copy, erase_graph
-from ..utils.symbol_utils import subs_idxc
 import torch.fx as fx
 import math
 from collections import defaultdict, deque, ChainMap
