+++ iree/turbine/kernel/wave/utils/mapping_utils.py
@@ -3,7 +3,7 @@
 # See https://llvm.org/LICENSE.txt for license information.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-from .symbol_utils import IndexExpr, IndexSymbol, subs_idxc
+from ..._support.indexing import IndexExpr, IndexSymbol, subs_idxc
 from ...lang.wave_types import IndexMapping
 import sympy
